refactor: useSyncExternalStore 대신 useRecruitInfo 직접 호출로 변경
